Move tree-item-base-vm.js to folder for "tree" components
