+++ src/ggrc/assets/javascripts/components/tree/tests/tree-item-base-vm_spec.js
