+++ src/ggrc/assets/javascripts/components/tree/tree-item-base-vm.js
